add grammar functional
