+++ admin_panel.py
@@ -79,16 +79,22 @@ async def add_document(message: Message, state: FSMContext):
     data = await state.get_data()
     hw: HW = data.get("hw")
     hw.file = document
-    await message.answer(f"Пришлите файл ответов или /cancel")
-    await NewHW.Answer.set()
     await state.update_data(hw=hw)
+    if hw.type != 'Grammar':
+        await message.answer(f"Пришлите файл ответов или /cancel")
+        await NewHW.Answer.set()
+    else:
+        text = f"Название: {hw.title}\nОписание: {hw.description}"
+        await message.answer_document(document=hw.file, caption=text)
+        await message.answer("Подтверждаете? Нажмите /cancel чтобы отменить", reply_markup=confirm_menu)
+        await NewHW.Confirm.set()
 
 
 @dp.message_handler(user_id=admin_id, state=NewHW.Answer, content_types=types.ContentType.DOCUMENT)
 async def add_answer_document(message: Message, state: FSMContext):
-    document = message.document.file_id
     data = await state.get_data()
     hw: HW = data.get("hw")
+    document = message.document.file_id
     hw.answer = document
     text = f"Название: {hw.title}\nОписание: {hw.description}"
     await message.answer_document(document=hw.file, caption=text)
