--- bd/db.html
@@ -1,12 +1,0 @@
-<!--[if IE]><meta http-equiv="X-UA-Compatible" content="IE=5,IE=9" ><![endif]-->
-<!DOCTYPE html>
-<html>
-<head>
-    <title>db.html</title>
-    <meta charset="utf-8"/>
-</head>
-<body>
-    <div class="mxgraph" style="max-width:100%;border:1px solid transparent;" data-mxgraph="{&quot;highlight&quot;:&quot;#0000ff&quot;,&quot;nav&quot;:true,&quot;resize&quot;:true,&quot;xml&quot;:&quot;&lt;mxfile host=\&quot;app.diagrams.net\&quot; modified=\&quot;2020-05-14T19:38:37.998Z\&quot; agent=\&quot;5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/81.0.4044.138 Safari/537.36\&quot; etag=\&quot;XJzavPkQVf5L6KNUiE8M\&quot; version=\&quot;13.1.1\&quot; type=\&quot;device\&quot;&gt;&lt;diagram id=\&quot;SRsH347r5aNN8QSTXPxD\&quot; name=\&quot;Page-1\&quot;&gt;7Vrbbts4EP0aA7sPuzAly3YeYzttduECixpF26eClWiJDUVqyVFs9+uXlEjLMpU0KXxZFAKMQHPE4e0cToZjD8J5vn0rcZG9Ewlhg2CYbAfhYhAE4yjQfw2wq4EwGtZAKmlSQ6gBVvQ7saBrVtKEqFZDEIIBLdpgLDgnMbQwLKXYtJutBWuPWuCUeMAqxsxHP9IEshqdBpMGvyc0zdzIaHxTv8mxa2xXojKciM0BFN4NwrkUAuqnfDsnzOyd25fa780Tb/cTk4TDSxwKMubZfSLV32KElvcflm+/yT+CkZ0c7NyKSaI3wJpCQiZSwTG7a9CZFCVPiOl2qK2mzVKIQoNIg98IwM6yiUsQGsogZ/Yt2VL4dPD82XT1Z2Stxdb2XBk7Z3CQu0+HxoGXMRu3ynJ+9frMop7cNwspUcqYPLNZU6s/LFMCz7RDN3t69bEgIid6QtpREoaBPrYngq1A0327hkP9YGl8BaXjut9HzEo70gdFpEez2tCcYW74XAsOjnGzZXFGWbLEO1GaVSrA8YOzZpmQ9Ltujx2X+rUEy3QwNr1RxuaCCakBLqoBGqeV6cwOI4nSbv84ItAR9A5vWw2XWIGboGAMF4p+raZsHHPNCeUzASDyFvGPRALZPk+9z5R1CO3RtbFrZM1NEwiQw7LDIBCOnubWDvZeByvMU72A/Wij9mgTf7Rxx2DBuD0WZkAkx0Bm5pQqT077Vf68wiaewkzz28E8GNwO/+JAUqM33fdwRf4tCdeH6lh9mhOoxCPFAzlSS4eAMKMp1yYja+NmSKU6Rt9aGEzcmakCx5Sny6rNYtQg7+1uGUho3zWr4nBGk4TwKp4BBvx1r/9CUA7VdkYz/dGbPjdxJlqYRUUz1Nj6Y5pLmAuu14JppSuilbohRq0vE+HTx9hX5q7N+o+keKyOk0WZqaeBUkeZL11C6Kk/KfUuo7oa9Qh53K9Lxr5wnJM9+yuQevt/Q8Ph770CTqyAyfSFCpieSQE3nYe/i/+op//k9KPhS4P/ufhHgSeA+499hvkTGeYI9Slmp8LCPsc8f5ypz/H/N8lEI08FQIH1WcZF6L9+ohl59CdExZIWQAU/FkHQi+AMIrh6ron8epYmro/2J2f6+mllEHmsXrAY7SrOrypGo6sVo12G9ONq9LSb9stUo5FfLFyYE9hfFl5/WTiuRyPUcVsIu24L4+DXvi341cj+tnD6/x+TV98WusR4vnTRL0spKPWGQl+WvgD/XdeFi/Lvvmg/4D/TXW+EfOgFcAEBdF0VOgVwtvzRL0uqMo6JUuuSGf6D4UwIRjDvyT81+Z23h8uy738tpdOwB1UT35/6cxHvKtFnIF6bze+j6mSx+ZFZePcf&lt;/diagram&gt;&lt;/mxfile&gt;&quot;,&quot;toolbar&quot;:&quot;pages zoom layers lightbox&quot;,&quot;page&quot;:0}"></div>
-        <script type="text/javascript" src="https://app.diagrams.net/js/viewer.min.js"></script>
-    </body>
-</html>
