+++ handlers.py
@@ -12,7 +12,7 @@ from auto_check import open_file
 from load_all import bot
 from pic_compare import compare_picture
 from python_check import compare_files
-from test import check_text
+from grammatic import check_text
 
 
 db = database.DBCommands()
@@ -83,7 +83,7 @@ async def enter_price(message: Message, state: FSMContext):
         answer_file = await bot.get_file(file_id=hw.answer)
         answer: io.BytesIO = await bot.download_file(answer_file.file_path)
     else:
-        answer = done.answer
+        answer = hw.answer
     test_file = await bot.get_file(file_id=done.answer)
     test: io.BytesIO = await bot.download_file(test_file.file_path)
     if hw.type == 'Test':
