--- text_file/answer.txt
@@ -1,5 +1,0 @@
-1. кошка
-2. стул
-3. собака
-4. bus
-5. floor
