--- text_file/hw_test.txt
@@ -1,5 +1,0 @@
-1. кошка
-2. стул
-3. собака
-4. car
-5. floor
