--- text_file/text_en.txt
@@ -1,9 +1,0 @@
-What makes every American a typical one is a desire to get a well-paid job that will cover their credit card. A credit card is an indispensable part of life in America. In other words, any American knows that how he or she handles their credit card or cards, either will help them or haunt them for years.
-
-In the U.S. getting a card isn’t as hard as it used to be. Some companies now mail applications to high school students, rather than waiting for them to get into college. At their best credit cards allow their owners to reserve hotel rooms, rent cars and finance larger purchases over several months. At their worst, cards allow people with poor money management skills to get into a high-interest debt.
-
-For those who are deep in credit card debt, there are some Credit Services agencies that offer anyone in America both online or telephone, and face-to-face counseling. The counselors may propose a debt management plan. The plan includes lower interest payments for clients and setting up a pay off timetable. The agencies’ average client makes about $32,000 a year and has more than $16,000 debt in the program. Usually the agencies charge zero up to $30 monthly to manage the debt service plan.
-
-Credit Services counselors advise, that if an American gets divorced, he or she may eliminate all joint debts by paying them off or transferring debt into a one-name account. Even in happy marriages, the agency advises that the husband and wife should each have a credit card in their name only to establish separate credit histories.
-
-Once debts have been repaid, an American can re-establish his/her good credit by applying for a secured credit card and paying the balance off regularly. A store credit card usually is the next step, followed by applying for a major credit card such as a Visa or MasterCard.
