--- text_file/text_ru.txt
@@ -1,33 +1,0 @@
-В этот вечер Пьер поехал к Ростовым, чтобы исполнить свое поручение. Наташа была в постели, граф был в клубе, и Пьер, передав письма Соне, пошел к Марье Дмитриевна, интересовавшейся узнать о том, как князь Андрей принял известие. Через десять минут Соня вошла к Марье Дмитриевне.
-— Наташа непременно хочет видеть графа Петра Кирилловича, — сказала она.
-— Да как же, к ней, что ли, его свести? Там у вас не прибрано, — сказала Марья Дмитриевна.
-— Нет, она оделась и вышла в гостиную, — сказала Соня.
-Марья Дмитриевна только пожала плечами.
-— Когда это графиня приедет, измучила меня совсем. Ты, смотри ж, не говори ей всего, — обратилась она к Пьеру. — И бранить-то ее духу не хватает, так жалка, так жалка!
-Наташа, исхудавшая, с бледным и строгим лицом (совсем не пристыженная, какою ее ожидал Пьер), стояла посередине гостиной. Когда Пьер показался в двери, она заторопилась, очевидно в нерешительности, подойти ли к нему, или подождать его.
-Пьер поспешно подошел к ней. Он думал, что она ему, как всегда, подаст руку; но она, близко подойдя к нему, остановилась, тяжело дыша и безжизненно опустив руки, совершенно в той же позе, в которой она выходила на середину залы, чтобы петь, но совсем с другим выражением.
-— Петр Кирилыч, — начала она быстро говорить, — князь Болконский был вам друг, он и есть вам друг, — поправилась она (ей казалось, что все только было и что теперь все другое). — Он говорил мне тогда, чтоб обратиться к вам...
-Пьер молча сопел носом, глядя на нее. Он до сих пор в душе своей упрекал и старался презирать ее; но теперь ему сделалось так жалко ее, что в душе его не было места упреку.
-— Он теперь здесь, скажите ему... чтоб он прост... простил меня. — Она остановилась и еще чаще стала дышать, но не плакала.
-— Да... я скажу ему, — говорил Пьер, — но... — Он не знал, что сказать.
-Наташа, видимо, испугалась той мысли, которая могла прийти Пьеру.
-— Нет, я знаю, что все кончено, — сказала она поспешно. — Нет, это не может быть никогда. Меня мучает только зло, которое я ему сделала. Скажите только ему, что я прошу его простить, простить, простить меня за все... — Она затряслась всем телом и села на стул.
-Еще никогда не испытанное чувство жалости переполнило душу Пьера.
-— Я скажу ему, я все еще раз скажу ему, — сказал Пьер, — но... я бы желал знать одно...
-«Что знать?» — спросил взгляд Наташи.
-— Я бы желал знать, любили ли вы... — Пьер не знал, как назвать Анатоля, и покраснел при мысли о нем, — любили ли вы этого дурного человека?
-— Не называйте его дурным, — сказала Наташа. — Но я ничего, ничего не знаю... — Она опять заплакала.
-И еще больше чувство жалости, нежности и любви охватило Пьера. Он слышал, как под очками его текли слезы, и надеялся, что их не заметят.
-— Не будем больше говорить, мой друг, — сказал Пьер.
-Так странно вдруг для Наташи показался этот его кроткий, нежный, задушевный голос.
-— Не будем говорить, мой друг, я все скажу ему; но об одном прошу вас — считайте меня своим другом, и ежели вам нужна помощь, совет, просто нужно будет излить свою душу кому-нибудь — не теперь, а когда у вас ясно будет в душе, — вспомните обо мне. — Он взял и поцеловал ее руку. — Я счастлив буду, ежели в состоянии буду... — Пьер смутился.
-— Не говорите со мной так: я не стою этого! — вскрикнула Наташа и хотела уйти из комнаты, но Пьер удержал ее за руку. Он знал, что ему нужно что-то еще сказать ей. Но когда он сказал это, он удивился сам своим словам.
-— Перестаньте, перестаньте, вся жизнь впереди для вас, — сказал он ей.
-— Для меня? Нет! Для меня все пропало, — сказала она со стыдом и самоунижением.
-— Все пропало? — повторил он. — Ежели бы я был не я, а красивейший, умнейший и лучший человек в мире и был бы свободен, я бы сию минуту на коленях просил руки и любви вашей.
-Наташа в первый раз после многих дней заплакала слезами благодарности и умиления и, взглянув на Пьера, вышла из комнаты.
-Пьер тоже вслед за нею почти выбежал в переднюю, удерживая слезы умиления и счастья, давившие его горло, не попадая в рукава, надел шубу и сел в сани.
-— Теперь куда прикажете? — спросил кучер.
-«Куда? — спросил себя Пьер. — Куда же можно ехать теперь? Неужели в клуб или в гости?» Все люди казались так жалки, так бедны в сравнении с тем чувством умиления и любви, которое он испытывал; в сравнении с тем размягченным, благодарным взглядом, которым она последний раз из-за слез взглянула на него.
-— Домой, — сказал Пьер, несмотря на десять градусов мороза распахивая медвежью шубу на своей широкой, радостно дышавшей груди.
-Было морозно и ясно. Над грязными, полутемными улицами, над черными крышами стояло темное звездное небо. Пьер, только глядя на небо, не чувствовал оскорбительной низости всего земного в сравнении с высотою, на которой находилась его душа. При въезде на Арбатскую площадь огромное пространство звездного темного неба открылось глазам Пьера. Почти в середине этого неба над Пречистенским бульваром, окруженная, обсыпанная со всех сторон звездами, но отличаясь от всех близостью к земле, белым светом и длинным, поднятым кверху хвостом, стояла огромная яркая комета 1812-го года, та самая комета, которая предвещала, как говорили, всякие ужасы и конец света. Но в Пьере светлая звезда эта с длинным лучистым хвостом не возбуждала никакого страшного чувства. Напротив, Пьер радостно, мокрыми от слез глазами смотрел на эту светлую звезду, которая как будто, с невыразимой быстротой пролетев неизмеримые пространства по параболической линии, вдруг, как вонзившаяся стрела в землю, влепилась тут в одно избранное ею место на черном небе и остановилась, энергично подняв кверху хвост, светясь и играя своим белым светом между бесчисленными другими мерцающими звездами. Пьеру казалось, что эта звезда вполне отвечала тому, что было в его расцветшей к новой жизни, размягченной и ободренной душе.
